project is ready to style.

+++ delete_confirmed.php
@@ -2,7 +2,7 @@
 require 'database/database.php';
 $pdo = db_connect();
 $banner = '';
-if ($_SERVER['REQUEST_METHOD'] == "POST" || isset($_POST['delete'])) {
+if ($_SERVER['REQUEST_METHOD'] == "POST" && isset($_POST['delete'])) {
     if (isset($_POST['user']) && isset($_POST['password'])) {
         delete_scores($_POST['user']);
         delete_user($_POST['user']);
@@ -11,3 +11,4 @@ if ($_SERVER['REQUEST_METHOD'] == "POST" || isset($_POST['delete'])) {
     }
 }
 include 'redirect.php';
+?>
